Update README.md
A deep learning and natural language processing project was developed to classify text sentiment as negative or positive. The project concluded with the creation of a user-friendly app where users can input their sentiments. A web app was built using Streamlit to deploy the model. Git was used to track changes in the app and update files in the GitHub repository. Docker was utilized to package all necessary dependencies and set instructions for deploying the model efficiently. Finally, the app was deployed on Render cloud, ensuring a smooth and scalable deployment process, making the sentiment analysis model accessible to users.